workspace/bots/ryan-bank/requirements.txt
discord.py>=2.3.0
apscheduler>=3.10.0
aiosqlite>=0.19.0
